* Fix GCC 3.4.5 compile error. * Fix int to float warning. * Remove FeatureDef destructor.
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@4518 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Sim/Weapons/Cannon.cpp
@@ -182,10 +182,10 @@ void CCannon::Fire(void)
 	if(weaponDef->flighttime > 0) {
 		ttl = weaponDef->flighttime;
 	} else if(selfExplode) {
-		ttl=(int)(predict+gs->randFloat()*2.5f-0.5f);
+		ttl = (int)(predict + gs->randFloat() * 2.5f - 0.5f);
 	} else if((weaponDef->groundBounce || weaponDef->waterBounce)
 			&& weaponDef->numBounce > 0) {
-		ttl = predict * (1 + weaponDef->numBounce * weaponDef->bounceRebound);
+		ttl = (int)(predict * (1 + weaponDef->numBounce * weaponDef->bounceRebound));
 	} else {
 		ttl=predict*2;
 	}
